Extract lean kirra-trajectory checker crate
