+++ crates/kirra-mick/examples/mick_sidewalk_brain.rs
@@ -15,9 +15,9 @@ use kirra_planner::{
     plan_for_intent, EgoState, GeometricPlanner, GeometricPlannerConfig, Goal, LlmBrain, MickBrain,
     MockModel, ObjectView, PlanInput, Pose, ProposalKind, WorldContext,
 };
-use kirra_ros2_adapter::corridor::{MockCorridorSource, Point};
-use kirra_ros2_adapter::state::{PerceivedObject, TrajectoryVerdict};
-use kirra_ros2_adapter::{validate_trajectory_slow, VehicleConfig};
+use kirra_trajectory::corridor::{MockCorridorSource, Point};
+use kirra_trajectory::state::{PerceivedObject, TrajectoryVerdict};
+use kirra_trajectory::{validate_trajectory_slow, VehicleConfig};
 
 const EGO_X: f64 = 2.0;
