+++ crates/kirra-planner/tests/adversarial_doer_bounded_by_kirra.rs
@@ -23,9 +23,9 @@ use kirra_planner::{
     plan_for_intent, EgoState, GeometricPlanner, Goal, MickIntent, PlanInput, PlanOutput, Planner,
     Pose, ProposalKind, TrajectoryPoint,
 };
-use kirra_ros2_adapter::corridor::{CorridorSource, MockCorridorSource, Point};
-use kirra_ros2_adapter::state::{PerceivedObject, TrajectoryVerdict};
-use kirra_ros2_adapter::{validate_trajectory_slow, VehicleConfig};
+use kirra_trajectory::corridor::{CorridorSource, MockCorridorSource, Point};
+use kirra_trajectory::state::{PerceivedObject, TrajectoryVerdict};
+use kirra_trajectory::{validate_trajectory_slow, VehicleConfig};
 use kirra_core::FleetPosture;
 
 /// KIRRA refuses to certify a trajectory longer than this many poses — a WCET bound
