+++ crates/kirra-planner/tests/learned_doer_invariance_capstone.rs
@@ -27,9 +27,9 @@ use kirra_planner::{
     PlanOutput, Planner, Pose, ProposalKind, Teacher, TrajectoryPoint,
 };
 use kirra_core::FleetPosture;
-use kirra_ros2_adapter::corridor::{CorridorSource, MockCorridorSource, Point};
-use kirra_ros2_adapter::state::{PerceivedObject, TrajectoryVerdict};
-use kirra_ros2_adapter::{validate_trajectory_slow, VehicleConfig};
+use kirra_trajectory::corridor::{CorridorSource, MockCorridorSource, Point};
+use kirra_trajectory::state::{PerceivedObject, TrajectoryVerdict};
+use kirra_trajectory::{validate_trajectory_slow, VehicleConfig};
 
 const SEED: u64 = 0xC0FFEE;
 /// The stopped car sits here, in-lane. A trajectory whose furthest reach exceeds this drives INTO
