+++ crates/kirra-planner/tests/turn_gap_acceptance.rs
@@ -15,9 +15,9 @@ use kirra_planner::{
     plan_for_intent, EgoState, FleetPosture, GeometricPlanner, Goal, Lane, LaneEdge, LaneGraph,
     LineType, MickIntent, PlanInput, PlanOutput, Pose, ProposalKind, TrajectoryVerdict, TurnDirection,
 };
-use kirra_ros2_adapter::corridor::Point;
-use kirra_ros2_adapter::state::PerceivedObject;
-use kirra_ros2_adapter::{validate_trajectory_slow, VehicleConfig};
+use kirra_trajectory::corridor::Point;
+use kirra_trajectory::state::PerceivedObject;
+use kirra_trajectory::{validate_trajectory_slow, VehicleConfig};
 
 const R: f64 = 12.0;
 const HALF: f64 = 3.0;
