+++ crates/kirra-ros2-adapter/Cargo.toml
@@ -94,6 +94,10 @@ smallvec = "1"
 # this severs the adapter→SDK edge and, transitively, the planner's last SDK pull.
 kirra-core        = { path = "../kirra-core" }
 parko-core        = { path = "../../parko/crates/parko-core" }
+# R1: the lean trajectory CHECKER (validation / prediction / perception_redundancy /
+# config / AcceptedTrajectory+EgoOdom). The adapter re-exports its modules so the
+# ros2/lanelet2 node code keeps using `crate::{validation,prediction,...}` unchanged.
+kirra-trajectory  = { path = "../kirra-trajectory" }
 
 # ROS 2 bindings — feature-gated. Pinned (`=`) because r2r explicitly does
 # not guarantee API stability; a deliberate bump is a separate maintenance
