+++ crates/kirra-trajectory/src/config.rs
