+++ crates/kirra-trajectory/src/perception_redundancy.rs
