+++ crates/kirra-trajectory/src/validation.rs
